Ensure datalogger version isn't float dtype in read result
Currently (but maybe not in pandas 3) the replace does result in a
dtype change if all values are "-9.000".

We can't cast to `str` since that converts `np.nan` to 'nan'.

Even without this change, when combining frames later,
it will become object dtype anyway as long as there is at least one
non-missing datalogger version entry in at least one frame.